Refactor event handling structures and enums
- Changed `EventType` enum from `int` to `byte` for memory efficiency.
- Added `InputEvent` struct to represent input events with relevant fields.
- Introduced `KeyButMask` enum for key and button masks as `ushort`.
- Added `MotionEvent` struct for motion events with coordinate properties.
- Redefined `XEvent` struct using `FieldOffset` for a union-like layout.
- Removed folder structure for `Models\Event` in the project file.

+++ Src/Models/Event/EventType.cs
@@ -1,6 +1,6 @@
 ﻿namespace Src.Models.Event;
 
-public enum EventType : int
+public enum EventType : byte
 {
     KeyPress = 2,
     KeyRelease = 3,
